requirements.txt
streamlit
langchain
langchain-core
langchain-openai
openai
python-dotenv
